+++ popular_proposal/apps.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class IniciativaPopularConfig(AppConfig):
-    name = 'iniciativa_popular'
+class PopularProposalConfig(AppConfig):
+    name = 'popular_proposal'
