+++ votainteligente/votainteligente_settings.py
@@ -45,7 +45,7 @@ INSTALLED_APPS = (
     'preguntales',
     # Uncomment the next line to enable admin documentation:
     #'django.contrib.admindocs',
-    'iniciativa_popular',
+    'popular_proposal',
 )
